--- Modification of function pnprint in PNLuaGameUtil.cpp. pnprint is now "pnint pnprint(const pnchar* format, int where = 0)"
where indicates where the infos have to be printed.

where = 0 stands for the console
where = 1 stands for the GUI console
where = 2 strands for writing in the lua debug log file ("pnscript.log")

default value is 0.

---
Two functions added to attach scripts to entities in the game, loadScript and unloadScript in PNGameMap.
Functions are commented for the moment (until names and role are defined more precisely).