First big update on the new "2011 port" branch.
* VCPP10 project created
* Boost lib update
* CEGUI updated to v0.75 

Holy crap! Crazy Eddy sure worked in 6 years! The guy changed everything... 
A lot of refactoring has been done, pnrenderer and pneditor work now, but pncore is still a mess and needs a lot of work (just to compile...).

More to come this summer, with hopefully a working program within september.

N.B: non-free data not committed at the moment. I can share them on a simple demand (without it the editor is just useless).
Linux Build is certainly out of sync but who plays on Linux anyway ? ;)
If someone wants to maintain it, feel free...